nina-api-client/pyproject.toml
[project]
name = "bundesamt-f-ü-r-bev-ö-lkerungsschutz-nina-api-client"
version = "1.0.0"
description = "A client library for accessing Bundesamt für Bevölkerungsschutz: NINA API"
authors = []
requires-python = "~=3.9"
readme = "README.md"
dependencies = [
    "httpx>=0.23.0,<0.29.0",
    "attrs>=22.2.0",
    "python-dateutil>=2.8.0,<3",
]

[tool.uv.build-backend]
module-name = "bundesamt_f_ü_r_bev_ö_lkerungsschutz_nina_api_client"
module-root = ""
data = [
    "CHANGELOG.md",
]

[build-system]
requires = ["uv_build>=0.8.0,<0.9.0"]
build-backend = "uv_build"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["F", "I", "UP"]
